Drop react-codemirror in favor of our own wrapper
We need to be able to subscribe to events before codeMirror renders and
react-codemirror doesn't let us do that. I tried adding it directly to
react-codemirror but I couldn't even get it to build properly. It also
has a bit of cruft that we don't need and this way we'll get true strong
typing