Check out this cool disabled state which doesn't work on Windows
Toggling this state works great but for some weird reason that I suspect
is a bug in Chrome it doesn't apply the opacity as long as the
.commit-button style has `overflow: hidden set` :rage:

Other properties work fine, like setting display: none or a different
background color.

I'm committing this for documentation and then implementing a
workaround.