Turns out this can have unintended consequences
By unsetting the HOME environment variable we essentially make git only read configuration from the repository directory and the system, skipping the global (home config).

This does very little to prevent what it intends to prevent, which is to ensure that our network operations don't end up invoking a credential helper.

If a credential helper is defined in the repository config that will still be used.

This also prevents users from setting using proxies or other configuration options that should apply globally for network operations. Most notably this prevents Git from reading user.name and user.email which is necessary for example for a pull operation where the pull ends up creating a merge commit.

Removing and figuring out a better approach