port registry usage to use PowerShell API
This circumvents the group policy which restricts
apps writing to the registry, which is what reg.exe
does despite it's best intentions. We can't fix that
here, so instead we're going to use the PowerShell API
for interacting with the registry, which has it's own
quirks. Because everything is terrible.