Remove O(N^2) loop in getStatus
For each item returned from `git status` we were doing an O(N) lookup (using findIndex) on the array to see if we had already added a file with that path. This change turns that array into a map keyed on the path so that we can perform O(1) lookups instead.

Measurements of a pathological case (10k files) show great improvement:

Before : 1519ms
After: 95.0ms