Publish-Repository no longer refreshes orgs when clicking current tab
Previously the organizations would be refreshed if the user clicked one of the tabs. It did not matter if it was the current tab or a different tab.

The comment indicates the intention was to refresh when changing to a different tab.

It will now only run when the tab is being changed to a different one and will not redo the work when staying on the same tab.