Prevent more top spacing from happening
Changing the `<p>` to a `<div>` removes the `margin-top` that gets added. I also updated the classname from `merge-info` to `merge-message` to be more consistent with the rest of the file. I style both class as the same.

We should consider settling on one classname to rule all merge status areas.