Repository branch names can no longer start with +
Previously the regular expression would check if the name began with a - and would replace that character with an empty character.

It has been updated to do this for both - and +.

This done because creating a new branch beginning with the character + results in a fatal error mentioning invalid refspec.