DestinationExistsError cleared if on default path and url
If the user is on the Clone-Repository page and has not changed the default path or url yet, a warning can still be generated if the path is validated.

This can be intimidating to see a red error when the user has not changed anything.

So now, if a DestinationExists error is displayed when the user is on the default path and url, it will be cleared. Otherwise the path will be validated on focus like normal.