Disable Commit Button when Summary is only spaces
Previously if a commit summary was only a single or more spaces and no other characters an error would be generated on commit button press and the commit would not be created.

Since this is not an ideal user experience, it has been changed to grey out the button when the commit summary is only spaces. It will re-enable when the commit summary contains at least one non-space character.