Renamed pull-request number variable to pullRequestNumber
Previously this variable name was 'number' which causes linting issues due to sharing the name of a type.

To get around this, the constructor initialized it with an argument named 'number_`

However, when switching to parameter properties the name 'number' is not allowed.

To prevent this issue the variable has been renamed to pullRequestNumber.