Separated Constructor and Class documentation
Previously the "A branch as loaded from Git" comment was for the Branch class. This was mistakenly combined with the constructor documentation, but still left in the class documentation spot.

The class comment has been restored to it's original form, and the constructor documentation has been moved to it's appropriate location.