Added discardingAllChanges boolean to DiscardChanges Popup Type
Previously the Discard Changes dialog did not have a way to tell if it is discarding 'all' changes. Discarding multiple changes at once used the same discardAllChanges functions therefore could not be distinguished.

This boolean allows passing that information.

This will eventually be used to allow the Discard Changes dialog to change it's text per each case.