Change display text when discarding all changes
Previously the distinction between discarding all changes and a single/several changes was not distinct.

This could lead to confusion when a user accidentally selects "Discard All Changes" instead of "Discard Change/s".

Now it is more clear by updating the dialog title and button text to clearly display that all changes will be discarded if they confirm the dialog.