Destination Already Exists error no longer shows for empty folders
Previously if a user created a folder they wanted to clone into, they would be presented with a destination already exists error blocking them from cloning.

This is because the function only detects if the folder exists, and does not care if the folder is empty.

Using isPathEmptyFolder(string) should allow them to clone if the folder does not contain any files already.