Display changed name warning when name has been sanitized
Repository name sanitization already occurred in the Create Repository dialog. However, when publishing the repository, the user is given another opportunity to change the name. The new name was not being sanitized, which allowed the user to enter unexpected and unintended characters such as emojis. Now the name will be sanitized in the Publish Repository dialog as well, however, a warning will display to show the user when the name is being changed and what the sanitized name is.