Pass initiateSyncOnCommit setting value down to commit-message.tsx
In order to optionally sync after a commit, the commit button function must know whether or not the setting has been enabled by the user.

If it is enabled, it will then sync on commit. But this feature is disabled by default, and must be able to skip this if it is not enabled.