Pull then Push if willInitiateSyncOnCommit is enabled by user
If the user has enabled the option to sync after creating a commit then after pressing the Commit button and creating a commit it will attempt to pull to get any commits preventing a push, then pushing the new commit to the remote repo.

However, if errors occur during the pull/push process the user will then have to handle them on on their own and manually push afterwards.