Pass remoteName from ui/App down to commit-message
Commit-message needs to be able to know whether or not the current branch it is creating a commit for has a remote.

This is because a feature is being implemented to allow immediate pushing after creating a commit. But this should not occur if the branch does not have a remote to push to.