Publish Repository now publishes with the sanitized name
Previously a warning was added to show to the user what their entered repository name will be changed to if it needs to be sanitized of illegal characters.

Now, the Publish Repository dialog will sanitize use the sanitized name during the publishing process. Previously, it displayed a warning but used the name as the user entered it.

In order to do this, a few changes needed to be made. A new private variable is used to store the name as the user entered it. This allows the program to know when the name has been changed in sanitization. If it has it will display a warning.

However, the name that is used is a sanitized version that is updated to the settings state.

Additionally, the onNameChanged function also updates the new name variable before updating the sanitized name.