Added Oversized-Files-Warning Popup Dialog
This is a new dialog that has been created to display a list of files that are over 100 MB to the user when they are attempting to commit large files without having Git LFS set up.

It provides the warning and information on Git LFS and offers the opportunity to cancel the commit or commit anyway.