Separated Git LFS check from checkForLargeFiles()
Previously the checkForLargeFiles() function would return an empty array if GitLFS was supported. This was not an obvious behavior to expect, and did not make too much sense.

Instead, the function will expect that if the function is called you wish to always get the same functionality. This means that the check is expected to be run before deciding to call the function.