Dispatcher is no longer passed to OversizedFilesWarning Dialog
The OversizedFilesWarning dialog needs access to the dispatcher to be able to initiate a commit.

Previously commit-message passed the dispatcher in to be sent along to the dialog in App.tsx.

But since App.tsx already has access to the dispatcher it is sufficient to simply pass it in from App.tsx and avoid the pre-emptive passing.