Added Pull Request List Filter Text Persistence
Previously any entered filter text in the drop-down pull request list would persist between switching tabs of the dropdown, but if the drop-down foldout was closed and reopened, the filter text would not be saved.

This has been changed so that when a user enters a filter text, leaves the foldout and returns their filter will be remembered and still be shown in the drop-down.