Moved Branch and Pull Request Filter variables into RepositoryState
Previously there was a singular branchFilterText and pullRequestFilterText saved for the entire program, stored within AppState.

This was not ideal since a filter text would only make sense for the repository it was set for.

So these variables are now stored in the RepositoryState so that each repository will have it's own that will be shown in the filter forms.