Replaced doesPathContainFiles with validateEmptyFolder function
This new function returns Promise<null | Error>

This allows us to quickly check if the result is a clonable path by way of checking for a null result.

If there was an error generated that prevents returning the result, it will be returned as an Error type that can be caught allowing the caller to decide what to do from there.