Set Error state directly with validateEmptyFolder's result
Previously the result of validateEmptyFolder would potentially be an an error meant for the caller function to make use of in order to make decisions or throw a friendlier user facing error.

Now validateEmptyFolder has been updated to have it's errors be user ready, so we can now set the error value directly to the result of the validation function.

The result will either be null, indicating no issues and ready to clone, or an error, which will be displayed in the UI for the user to notice.