Added readMeExists state variable
This variable is used to track whether or not the directory the user is attempting to create a directory to contains a ReadMe.md file.

This is tracked so that the user can be warned if initializing the new repository with a readme will result in overwriting their existing file.