Added doesReadMeExist, and now updates readMeExists variable
This function returns a boolean depending on whether the directory the user has selected to create a repository into already contains a ReadMe.md file.

This is used to update the readMeExists variable on name change and on path change so that it can display a warning when there is a readme file that may be overwritten if the new repository is initialized with a readme file.

For a filename to return match, it must be "ReadMe.md" (with allowance for different capitalization).