Add support for WSL shell in Windows
The shell is invoked with `wsl.exe`, which uses whichever Linux
distribution the user has configured as their default. [1]

The option is presented to the user when the following conditions
are met:

1. `wsl.exe` and `wslconfig.exe` are both found in the system.
2. `wslconfig.exe /list` returns an exit code of zero.

I verified that `wslconfig.exe /list` returns a non-zero exit code
when the system has no Linux distributions installed. I decided to
avoid the registry for detection since that has been discouraged
by the WSL developers. [2]

Closes #4641

[1]: https://docs.microsoft.com/en-us/windows/wsl/wsl-config#wsl-and-wsl-command
[2]: https://github.com/Microsoft/WSL/issues/3515