Added onRowKeyDown prop to filter-list
Currently the filter list has an onRowKeyDown function that calls when a key is pressed while a filter list row is in focus.

This adds an optional prop that is called when this event occurs if a function is passed down. This allows the parent class to be able to respond to this event with it's own actions.