Cloning now occurs on Enter press with repository list item in focus
Previously if a user selected a repository from the repository list to clone and hit enter while the list row was in focus nothing would occur as a result.

Since the expectation is that on enter a clone would initiate for the selected repository row it has been updated to do this.