Clone Repository dialog can now display "No internet connection" warning
Previously if a list of repositories could not be loaded for a user it always say "Couldn't find any repositories" for that account.

This may imply to the user they may not have any repositories.

However, this may also be due to not having an internet connection when attempting to clone.

Therefore if the user is connected to a network then they will be shown the "Could not find any repositories" dialog. Otherwise they will see a new "No internet connection" warning which includes a link to refresh.