Added check for cloning being enabled before initiating clone on repo list enter
Previously checks were not done to ensure things were in a clone-able state before attempting to initiate cloning on repository list item enter presses.

Now there has been an added function that performs the same check as is done when enabling/disabling the Clone button and initiates a clone only if it passes the check.