Added retry-clone-dialog.tsx for handling clone git errors
Previously if a clone failed and returned a git error such as when failing due to connection issues, a git error was presented directly to the user.

These errors are not great for the user as they are intimidating and potentially confusing or vague.

This dialog will display to the user that cloning has failed and if possible, offer them the chance to retry cloning.