Update Clone Error Handler to accept any clone error with retry action
Previously this handler was designed to only work with the HostDown dugite clone error.

However, the purpose has shifted to make retrying a clone for any type of clone error that is thrown.

The function will escape by returning the error immediately if the retryAction (if any) is not a Clone retryAction.

This should allow throwing up the retry popup only when an error is a clone error and has the ability to retry.