Fix accidental refactoring bug
In 79a85a34a6e5bf2976510d9a6ebf02838686e46f we were shuffling some code around and in the proccess we inadvertently moved the check for whether the main window exists to after we've destroyed it and cleaned up the handle.

The end result being that we always assume that a crash is the result of a launch problem.