Recreate pullRequest table, change primary key
This rewrites the pullRequest table to remove the previous automatically incrementing primary key and replacing it with a compound index on base.repoId and pr number, the combination of which is unique.

This allows us to simplify a lot of our logic and brings the additional benefit that we can do bulk deletes and upserts without first having to query to see whether there's already a matching item in the table.