Added CodeRunner to the External Editor list
Added to CodeRunner: https://coderunnerapp.com/ to the list of External Editors.

The only section I was unsure about is the function getExecutableShim() section.  I attempted to follow the instructions until I got to https://github.com/desktop/desktop/blob/development/docs/technical/editor-integration.md#step-2-find-executable-to-launch, which was not very helpful so I just mirrored the majority of the other applications hoping this is what is needed.