Log `shell.openExternal` errors
Electron's shell.openExternal changed to become async so in order to ensure that errors don't get lost in the async void we need to explicitly catch them.

I'm choosing to log them rather than throw because surfacing the error to the user doesn't add any value. They'll notice that something is wrong by virtue of the fact that their browser didn't open correctly.

The logging will be helpful for us developers when we're trying to debug what went wrong.