Make app's version string selectable text instead of a special link
The version string in the About dialog looks like other links in the same dialog (blue, underline on hover), but is not really a hyperlink. Instead, clicking on it puts the version number to clipboard.

This change makes the version text selectable instead and relies on the user to select it and copy to clipboard manually if they so desire. It aims to make the About dialog a little more intuitive.