Rename getSelectedExternalEditor --> lookupSelectedExternalEditor
This might be just me, but when I think `get` I imagine a simple getter that looks up and returns a value. Whereas this function is a bit more complex -- it looks in local storage and then looks up available editors per platform, checking the cache first. The term `lookup` may not be the bet either, but it's the best I can think of.

I also wanted to do this because I am going to create an `updateSelectedExternalEditor` method, and it'd be weird to have `updateSelectedExternalEditor(this.getSelectedExternalEditor)`