I am so smart! S-M-R-T... I mean S-M-A-R-T!
In between Git 2.22.1[1] and 2.24.0[2] Git attempted to use a different method of clearing the last line in the terminal for non-dumb terminals. While that change was reverted in 2.24 we're currently on 2.23 and have to deal with it.

The side effects of this for us is that our progress parsing has stopped working and that Git error messages are littered with lines starting with "[K" which at best will be confusing to users.

Setting TERM to 'dumb' effectively reverts the behavior to what we're used to (see https://github.com/git/git/blob/a7312d1a28ff3ab0a5a5427b35f01d943103cba8/editor.c#L11-L15).

1. https://github.com/git/git/commit/5b12e3123b7b70e3875404a4ffe571ca079364fe#diff-cf9e2e00bf41dfd6ab5e3e60149405dd
2. https://github.com/git/git/commit/bbf47568ad7e91ab0962b314c054a2da03232c72#diff-cf9e2e00bf41dfd6ab5e3e60149405dd