Open in Command Prompt: Do use the absolute path of the found shell
We go out of our way to find the correct path of the shell, so we might just as well use it. This will also help with _not_ picking up a script in the current working directory on Windows: if executing, say, `pwsh` in a directory where a script `pwsh.bat` exists, that script is executed instead of the `pwsh.exe` on the PATH.

Please note that the parameters have to be quoted *explicitly* on Windows (this seems to be a quirk of `spawn()` when called with `shell: true`). But that comes with an extra complication: When opening a program whose path has spaces via `START`, we cannot simply quote it: according to `start /?`, if the first parameter is enclosed in double-quotes, it specifies a window title. So we're providing window titles for all the changed call sites so that the paths _can_ be quoted.