Header, License
Added Header to let users/ client to access the source code. 

also, I think we need to discuss about License. Shouldn‘t we change it to Open source instead of MIT. As previous version mentioned MIT License is not for GitHub. so temporary empty it, for us to discuss what we should adapt a suitable license for this GitHub Desktop.