Fix bug in fetchAll for non-array values returned from parsedResponse
The parsedResponse function just converts the received JSON directly to an object without doing any checking. This causes problems for fetchAll if the response ends up being a scalar object rather than an Array. As a result the call to `buf.push(...)` was failing.

Now we ensure that it is an array.