Collapse organisations in Clone Repositories window (#11908)
Make the organisations groups in the clone repos window collapsed by default. Only then load repositories for an organisation if the user expands it.
This has the advantage of making the clone window less cluttered and load faster in cases where there are a large number of repos in an organisation.

At the moment there are a couple of bits that still need some work:

  *  At the moment it requires the list each time you expand an organisation. As a result if you collapse then re-expand the organisation, it will do the query again (effectively a refresh). This could be optimised to not remove the entries when an org is collapsed.
  *  It currently only works for organisations with public membership of the user. I don't really know why this is. I'm using the `GET user/orgs` API query to fetch this list, but despite the fact that we should be authenticated, it is not returning organisations with private membership.