Correct type for IFilterListCollapsableGroup
This should be a direct combination of the `IFilterListGroup` and the type `T` as it must be able to behave like an item. Previously it was behaving as just the base `IFilterListItem`.

This has the added advantage that we can use the `icon` property from `ICloneableRepositoryListItem` to store our collapsed/expanded icon rather than fixing it in the filter list class.