Fix selection of organisation groups in clone repository window
We should be able to select the group headers using the keyboard to allow using space/enter to expand/collapse the group. This was not working because the group headers are organisations not repositories which means the logic was getting confused and reverting the selected item to null whenever an organisation was selected.

Now we allow the selected item to be either a repo or an org, and use careful type checking to work out which one we are looking at. If a repository, then we allow the option of cloning it, updating the local path as before. If an organisation, we don't allow cloning, but we do allow selecting.