Fix issue #13636 - Tooltip position is (0,0) if mouse is not moved
The tooltip position is referenced to mouse position set in mousemove event. If the tooltip is triggered by events other than mousemove (e.g. scroll the list with mousewheel or keyboard arrow keys, or click the Fetch Origin button without moving the mouse), the mouse position is not set (0,0).  The fix is to set the mouse position in mouseenter event as well