Add IntelliJ IDEA Comunity as supported editor
I added just the snap path for the community edition because it is the version I know where it is installed, since it is the one I have. 

Also, I'm not sure if community should be differentiated from ultimate, but I, for sure, would differentiate it from the educational edition.