Get rid of broken mouse over event in lists
The `focusOnHover` only comes into play when `selectOnHover` is set and the only component using that (now) is the `AutoCompletingTextInput`.

Due to a bug in the mouseOver event handler on the List component we would only invoke the `selectionChanged` event and not the `selectedRowChanged` event meaning that hovering wouldn't ever change the selection. The only thing it would do is scroll to partially visible rows which isn't a common behavior and not worth keeping all this logic (and a11y lint exception) around for.