Support preview for DDS images
Uses WebGL2 to render DDS images to a canvas, then extract that canvas image to a dataURL that can be passed into the existing logic to render image previews.

This also necessitates adding the rawContents (buffer) to the Image class so we can use that in the conversion later on. This is technically an optimization since we could decode the base64 contents, but that would introduce an additional (and significant) delay in the conversion that is unnecessary.