Center empty states in branches and pull requests container
When there are no branches or pull requests to show, the corresponding empty states are centered on the container with a fixed width of 365px.

For the branches empty state container, the fixed width replaces `flex: 1` to prevent it from filling the container and make the Create Branch button too wide.