Delete app-store-test.ts
AppStore tests are flakey and we've ended up disabling all but one of them. AppStore is just not set up for testing right now with it spawning unobservable async operations and setting up timers in its constructor. I don't want to invest time into fixing it and the test in question is checking whether we can select a repository or not and if we fail to catch that during manual testing we deserve what's coming for us.